requirements.txt
asgiref==3.8.1
Django==4.2.13
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
